requirements.txt
flask
flask_login
gitpython
requests
psycopg2
cryptography
waitress
celery
